test_basic_rerank_caching
